solution: day3 part2
Compared to the previous version, we store the steps needed to reach every
single point of each wire, so we will have a (point, steps1, steps2) tuple,
of which we have to search the minimum of (steps1 + steps2) and that's it.

The algorithm is essentially the same as before,
just with different data structures.