refact: make computer an object
There is too much state already and memory read and write operations need to be
in a central place because of the next feature: allocating more memory on demand.
Currently it would not have been possible, because every read and write
operation was handled at the place of running the instruction and would have
been a nightmare to include that feature everywhere.