feature: expand memory on-demand
when reading, we don't need to expand, just return 0
when writing, epxand with 0 with the minimum amount of memory needed for the new
address.

Now that we can read further, we can get param2 early, it will just read 0 when
we reach out of memory bounds, not blow up with index error.