refact: day7 run every permutation in a different goroutine
Start a goroutine for every phase setting, send the results through a channel
and collect the results from that channel to search for the maximum until the
channel is closed (which means there will be no more results).

Note this run the same amount of time than the previous version and it is more
complicated, it's just a learning material.