refact: make itertools.Permutations type-agnostic
It's pretty painful, because we have to use empty interface slice, but we don't
touch the elements inside it, so we can pass in any type really. The only
problem is that the list have to be passed before iterations and converted back
to the original types after we got a result.